--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": ["env"],
-  "plugins": [
-    "transform-object-assign",
-    "transform-es3-member-expression-literals",
-    "transform-es3-property-literals"
-  ]
-}
