+++ package.json
@@ -24,9 +24,10 @@
   "scripts": {
     "prepare": "husky install",
     "lint": "eslint '{src,test,lib}/**' --ext .js,.jsx,.ts,.tsx",
+    "check": "tsc",
     "prekarma": "webpack",
     "karma": "karma start",
-    "pretest": "npm run lint",
+    "pretest": "npm run check && npm run lint",
     "test": "npm run karma",
     "test:bs": "BS=true npm run karma",
     "prebuild": "npm run test",
@@ -37,16 +38,13 @@
     "es6-promise": "4.2.8",
     "js-cookie": "2.2.1",
     "jssha": "3.2.0",
+    "tslib": "2.2.0",
     "uuid": "8.3.2",
     "xr": "github:dhinus/xr"
   },
   "devDependencies": {
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
-    "babel-plugin-transform-es3-member-expression-literals": "6.22.0",
-    "babel-plugin-transform-es3-property-literals": "6.22.0",
-    "babel-plugin-transform-object-assign": "6.22.0",
-    "babel-preset-env": "1.7.0",
+    "@types/karma": "6.3.0",
+    "@types/node": "14.17.0",
     "chai": "4.3.4",
     "eslint": "7.26.0",
     "eslint-config-contactlab": "6.2.1",
@@ -64,6 +62,9 @@
     "pretty-quick": "3.1.0",
     "puppeteer": "9.1.1",
     "sinon": "10.0.0",
+    "ts-loader": "9.2.1",
+    "ts-node": "9.1.1",
+    "typescript": "4.2.4",
     "webpack": "5.37.1",
     "webpack-cli": "4.7.0",
     "webpack-merge": "5.7.3"
