+++ test/integration/customer.js
