+++ test/integration/load.js
