+++ tsconfig.build.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "target": "ES6",
+    "module": "ESNext"
+  },
+  "include": ["./src"]
+}
