+++ tsconfig.build.legacy.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "target": "ES5",
+    "module": "UMD"
+  },
+  "include": ["./src"]
+}
